created new AbstractServerTask class
