+++ src/main/groovy/net/wasdev/wlp/gradle/plugins/tasks/StopTask.groovy
@@ -17,7 +17,7 @@ package net.wasdev.wlp.gradle.plugins.tasks
 
 import org.gradle.api.tasks.TaskAction
 
-class StopTask extends AbstractTask {
+class StopTask extends AbstractServerTask {
 
     @TaskAction
     void stop() {
